second commit

+++ BakeryFinal/Model/DTO/OrderDTO.cs
@@ -2,7 +2,7 @@
 {
     public class OrderDTO
     {
-        public int TotalPrice { get; set; }
+        public float TotalPrice { get; set; }
         public int BakeryOfficeId { get; set; }
         public virtual ICollection<BreadOrderDTO> BreadOrder { get; set; } = null!;
         public string Status { get; set; }
